fix: Create separate reducer for last indexed value (#543)
* Added required new files

* Added new reducer for handling lastIndexed

* removed all comments

* suggested changes
