+++ amundsen_application/static/js/components/Footer/index.spec.tsx
@@ -85,6 +85,6 @@ describe('mapStateToProps', () => {
   });
 
   it('sets lastIndexed on the props', () => {
-    expect(result.lastIndexed).toEqual(globalState.tableMetadata.lastIndexed);
+    expect(result.lastIndexed).toEqual(globalState.lastIndexed.lastIndexed);
   });
 });
