+++ amundsen_application/static/js/ducks/tableMetadata/sagas.ts
@@ -11,15 +11,11 @@ import {
   getTableDescriptionSuccess,
   getColumnDescriptionFailure,
   getColumnDescriptionSuccess,
-  getLastIndexedFailure,
-  getLastIndexedSuccess,
   getPreviewDataFailure,
   getPreviewDataSuccess,
 } from './reducer';
 
 import {
-  GetLastIndexed,
-  GetLastIndexedRequest,
   GetPreviewData,
   GetPreviewDataRequest,
   GetTableData,
@@ -162,20 +158,6 @@ export function* updateColumnDescriptionWatcher(): SagaIterator {
     UpdateColumnDescription.REQUEST,
     updateColumnDescriptionWorker
   );
-}
-
-export function* getLastIndexedWorker(
-  action: GetLastIndexedRequest
-): SagaIterator {
-  try {
-    const lastIndexed = yield call(API.getLastIndexed);
-    yield put(getLastIndexedSuccess(lastIndexed));
-  } catch (e) {
-    yield put(getLastIndexedFailure());
-  }
-}
-export function* getLastIndexedWatcher(): SagaIterator {
-  yield takeEvery(GetLastIndexed.REQUEST, getLastIndexedWorker);
 }
 
 export function* getPreviewDataWorker(
